Rename incremental/incr_stage.py to incremental/stage.py
